Wrap nav in suspense
